ext/gd/config.m4: fix cross-compilation in PHP_GD_CHECK_FORMAT macro
The AC_RUN_IFELSE inside PHP_GD_CHECK_FORMAT was missing its third
argument, namely what to do when the user is cross-compiling. Its
absence is not immediately fatal, but it does cause ./configure to
bail out if the user attempts to cross-compile. We add a pessimistic
default (format not supported) for that case to allow ./configure to
proceed.