round: Bypass the precision logic when rounding to 0 places
Since GH-12220 the implementation of `php_round_helper()`, which performs
rounding to an integral value, is easy to verify for correctness up to the
floating point precision.

If rounding to 0 places is desired, i.e. the userland `round()` function is
called with `$precision = 0`, we bypass all logic for the decimal point
adjustment and instead directly call `php_round_helper()`.

This change fixes the remaining two cases of GH-12143 and likely guarantees
correct rounding for all possible inputs and `$precision = 0`.