Fix GH-12778: Windows IIS FastCGI Spin-Up Performance Slowdown
The issue isn't limited to Windows only, I can also see the slowdown on
my Linux machine. The cause of the slowdown is 26e424465c. In that
commit a locale reset routine was implemented that gets called on
startup.
The fix from that commit is only necessary when using readline
functionality, there are three places that is used:
  - phpdbg
  - with readline_cli
  - readline() php function
So we make sure we only reset the locale when one of these is used,
preventing the overhead on normal website-serving use-cases.