Use zval_try_get_long() in NumberFormatter::format()
To improve precision handling in cases where the number doesn’t fit into
a 32-bit or 64-bit integer. A float that doesn’t fit into a long will
produce a deprecation warning; for NumberFormatter::TYPE_INT32, a number
that doesn’t fit into 32 bits will additionally produce a ValueError.
(This is inconsistent, but I’m not sure how to do it better.)