ext/gd/config.m4: don't rename HAVE_GD_JPG
The new gd format detection macro accidentally renames HAVE_GD_JPG to
HAVE_GD_JPEG because gd uses "Jpeg" in its function names. But this
would break consumers for no reason. Instead we require the caller to
pass in the name of the constant explicitly, and when checking for
Jpeg support, we pass in HAVE_GD_JPG.

Thanks to Peter Kokot for the suggestion.