Enhance moving PHP code pages into huge pages
The former implementation of huge pages for PHP code segments may
fail to map in certain scenarios. For instance, if the initial
'r-x-' segment is not PHP, it will result in a failure to map into
huge pages. Consequently, the optimization for huge pages with PHP
code will no longer be effective.

This patch improves the implementation by accurately matching all
'r-x-' segments until it locates the PHP code segment. Subsequently,
it performs the necessary huge page mapping.

Signed-off-by: PeterYang12 <yuhan.yang@intel.com>
Reviewed-by: chen-hu-97 <hu1.chen@intel.com>