Fix GH-12850: DOMNameSpaceNode parent not unset when namespace gets unlinked
The problem is that the namespace nodes are exposed as fake nodes
because of their different underlying data structure, which means that
the parent link isn't actually updated when it gets removed; and we
don't have a simple way of tracking it.

So just perform a check for the node type and check whether the
namespace declaration associated with the fake node is still in
existence.