Fix creating a first-class callable from FFI
This causes a NULL pointer access because the scope field is being
accessed in `zend_closure_call_magic`. We get to that function because
of the ZEND_ACC_CALL_VIA_TRAMPOLINE fn_flag that is checked in
`zend_closure_from_frame`. The `zend_closure_from_frame` function does
not take into account the reserved data or the internal handler
function, causing us to end up in `zend_closure_call_magic` instead of
the FFI trampoline.

Get rid of the flag, and store the temporary internal function in
another reserved slot so we can clean that up at the end of the FFI
trampoline's execution.