Remove properties field from php_libxml_node_object
This shrinks the struct from 80 bytes to 64 bytes.
This was unused internally, I did not find users externally via GitHub
search.
The intention for this was that it could be used for attaching extra
data as a 3rd party to a node. However, there are better mechanisms for
that like using actual objects.